Start port to GTK using GIR
As the use of the tray icon in wxWidgets seems a bit problematic, as the icon
disappears, or is put as some kind of special window instead of the taskbar, I
presume in case the notification area hasn't started yet.

This, little, issues drove me to look a bit more to GTK. Benefits are it
supports more languages, and it's a native library. Downside could be it's
less portable, but the question is if it portable anyway at the moment, as
unison requires SSH, which is, according to their website, a bit problematic on
Windows.

And, I myself, wouldn't do much for it anyway, as I don't run it, and it would
be a big, and quite impossible (without running Windows), task to support it.

And, I've got to say, I quite like the API, at the moment. And, I'm happy that
it supports handling the default system icons, which makes things easier for me.
The result will be better then it was using wxWidgets, and I doubt if there are
some real downsides to the new approach, as, well, it seems to me like the only
downgrade is regarding Windows, and I don't care about that.