Show message when done
The conversion is now complete, except from the preferences editor, but I don't
care.

Note GObject.thread_init(). I've been debugging quite a long time before I
realized the only reason it didn't work was GTK.

Anyway, I'm happy with the result, and the GUI is completely separated with the
actual logic, which should make it quite easy to convert to Qt, if I finally
decide to do so.