Revert "Last sync should be increased by length of 1 tick"
This reverts commit b12c85f2a2303fd4c263194527ca55b5a39dac32.

The reason for this revert is that this commit contains code which
should never have been committed in the first place. I tried to use
pygtk, which unfortunately is not available for Python 3, as GTK 3.14
has virtually deprecated the whole GUI by deprecating StatusIcon. As
the trial with pygtk was unsuccessful, unisister will probably be
ported to Qt.